requirements.txt
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
numpy==1.26.4
pandas==2.2.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-telegram-bot==12.8
pytz==2024.1
six==1.16.0
tornado==6.4
tzdata==2024.1
